Python 3 Compatibility

+++ test-setup.sh
@@ -1,6 +1,11 @@
-mkdir -p setup-py-tst
+rm -rf build
 python setup.py build
-LIB=`find build -name pyca.so`
-LINK='setup-py-tst/pyca.so'
-rm $LINK
+LIB=`find build -name pyca*.so`
+LINK='pyca.so'
+if [ -L $LINK ]; then
+  rm $LINK
+fi
+if [ -z $LIB ]; then
+  exit
+fi
 ln -s $LIB $LINK
